Install on Windows as a PowerShell Module
Converts the ps1 script to a script modeule (psm1) and makes some
updates to the syntax and behavior to resolve bugs in the existing
script. This is not considered "final", as there is at least one item
case where it hasn't reach feature parity with the Linux script
$CLJ_CONFIG. A more thorough comparison of the two scripts needs to be
done to ensure they are equivalent.

The main benefit of installing as a module is that PowerShell provides
this native capability to package functionality and install it in common
process, which does not require adding items to the users PATH or
installing in a wider variety of folders. Down the line, if this were
published in the PowerShell Gallery the entire install process would be
"Install-Module ClojureTools" with a similar command to receive updates.

Main drawback is that it requires more direct use of PowerShell, since
no commands are placed on the traditional PATH. However, with PowerShell
becoming a bigger part of Windows dev usage, this may not be seen as a
significant drawback.

Signed-off-by: Alex Miller <alex.miller@cognitect.com>