Allow skipping parsing with --
This allows users to specify that they no longer want Clojure to try and
interpret the arguments.  This allows for tools which wrap tools.deps
directly (e.g. pack) to take arguments which are compatible with
clojure, e.g. `clojure -A:pack -- -A:kafka` would pass `-A:kafka` to
pack, as opposed to `-- -A:kafka`.

Signed-off-by: Alex Miller <alex.miller@cognitect.com>